Fix item list layout header rendering bug
