--- src/renderer/components/item-object-list/header-filters.tsx
@@ -1,17 +1,0 @@
-/**
- * Copyright (c) OpenLens Authors. All rights reserved.
- * Licensed under MIT License. See LICENSE in root directory for more information.
- */
-import { observer } from "mobx-react";
-import React from "react";
-import type { HeaderPlaceholders } from "./list-layout";
-
-export interface ItemListLayoutHeaderFiltersProps {
-  headerPlaceholders: HeaderPlaceholders
-}
-
-export const ItemListLayoutHeaderFilters = observer(({ headerPlaceholders }: ItemListLayoutHeaderFiltersProps) => (
-  <>
-    {headerPlaceholders.filters}
-  </>
-));
